convert facade to ts, add types
