--- src/components/pseudonym/index.js
@@ -1,5 +1,0 @@
-const pseudonymFacade = require('./uc/pseudonym.facade');
-
-module.exports = (app) => {
-	app.configure(pseudonymFacade);
-};
