+++ src/components/pseudonym/index.ts
@@ -1,0 +1,6 @@
+import { Application as FeathersApplication } from '@feathersjs/feathers';
+import pseudonymFacade from './uc/pseudonym.facade';
+
+export default (app: FeathersApplication): void => {
+	app.configure(pseudonymFacade);
+};
