--- src/components/pseudonym/uc/pseudonym.facade.js
@@ -1,9 +1,0 @@
-const pseudonymUC = require('./pseudonym.uc');
-
-const facade = {
-	deleteUserData: pseudonymUC.deleteUserData,
-};
-
-module.exports = function setupFacade(app) {
-	app.registerFacade('/pseudonym/v2', facade);
-};
