+++ src/components/pseudonym/uc/pseudonym.facade.ts
@@ -1,0 +1,10 @@
+import { Application as FeathersApplication } from '@feathersjs/feathers';
+import { deleteUserData } from './pseudonym.uc';
+
+const facade = {
+	deleteUserData,
+};
+
+export default (app: FeathersApplication): void => {
+	app.registerFacade('/pseudonym/v2', facade);
+};
