+++ types/index.d.ts
@@ -1,5 +1,8 @@
 import type { ObjectId as MongooseObjectId } from 'mongoose';
 
+/**
+ * we allow object id's to be a string or the mongoose object id.
+ */
 export type ObjectId = string | MongooseObjectId;
 
 export interface TrashBinResult<T> {
